Alterado int p/ byte

+++ Src/Controllers/SobreController.cs
@@ -34,8 +34,8 @@ namespace API_REST_The_Last_Of_Us.Controllers
 
             return (Status) switch
             {
-               (int)EnumUtils.StatusProc.Sucesso => new OkObjectResult(Json),
-               (int)EnumUtils.StatusProc.ErroServidor => null,
+               (byte)EnumUtils.StatusProc.Sucesso => new OkObjectResult(Json),
+               (byte)EnumUtils.StatusProc.ErroServidor => null,
                _ => throw new System.NotImplementedException()
             };
          }
